Sync with 23.08.0 dev

See merge request ndip/galaxy!57

+++ lib/galaxy/webapps/galaxy/api/job_files.py
@@ -97,14 +97,18 @@ class JobFilesAPIController(BaseGalaxyAPIController):
             )
             assert file_path.startswith(
                 upload_store
-            ), "Filename provided by nginx (%s) is not in correct" " directory (%s)" % (file_path, upload_store)
+            ), f"Filename provided by nginx ({file_path}) is not in correct directory ({upload_store})"
             input_file = open(file_path)
         else:
             input_file = payload.get("file", payload.get("__file", None)).file
         target_dir = os.path.dirname(path)
         util.safe_makedirs(target_dir)
         try:
-            shutil.move(input_file.name, path)
+            if os.path.exists(path):
+                with open(path, "ab") as dest:
+                    shutil.copyfileobj(open(input_file.name, "rb"), dest)
+            else:
+                shutil.move(input_file.name, path)
         finally:
             try:
                 input_file.close()
@@ -122,7 +126,7 @@ class JobFilesAPIController(BaseGalaxyAPIController):
 
         job_id = trans.security.decode_id(encoded_job_id)
         job_key = trans.security.encode_id(job_id, kind="jobs_files")
-        if not util.safe_str_cmp(kwargs["job_key"], job_key):
+        if not util.safe_str_cmp(str(kwargs["job_key"]), job_key):
             raise exceptions.ItemAccessibilityException("Invalid job_key supplied.")
 
         # Verify job is active. Don't update the contents of complete jobs.
@@ -142,15 +146,8 @@ class JobFilesAPIController(BaseGalaxyAPIController):
         https://gist.github.com/jmchilton/9103619.)
         """
         in_work_dir = self.__in_working_directory(job, path, trans.app)
-        allow_temp_dir_file = self.__is_allowed_temp_dir_file(trans.app, job, path)
-        if not in_work_dir and not allow_temp_dir_file and not self.__is_output_dataset_path(job, path):
+        if not in_work_dir and not self.__is_output_dataset_path(job, path):
             raise exceptions.ItemAccessibilityException("Job is not authorized to write to supplied path.")
-
-    def __is_allowed_temp_dir_file(self, app, job, path):
-        # grrr.. need to get away from new_file_path - these should be written
-        # to job working directory like metadata files.
-        in_temp_dir = util.in_directory(path, app.config.new_file_path)
-        return in_temp_dir and os.path.split(path)[-1].startswith("GALAXY_VERSION_")
 
     def __is_output_dataset_path(self, job, path):
         """Check if is an output path for this job or a file in the an
